exchange rate massive changes

+++ backend/models/payment.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Optional
 
 from pydantic import BaseModel, Field
@@ -9,12 +10,12 @@ class PaymentCreate(BaseModel):
     # The debtor (from_member_id) is the payer; the creditor (to_member_id) is the receiver.
     from_member_id: str
     to_member_id: str
-    amount: float = Field(gt=0, allow_inf_nan=False)
+    amount: Decimal = Field(gt=0, allow_inf_nan=False)
     note: Optional[str] = None
 
 
 class PaymentPatch(BaseModel):
     # Edit an existing payment (PATCH /trips/{id}/payments/{pid}). Only amount/note are mutable;
     # the direction (from/to members) is fixed. Both fields optional so either can be sent alone.
-    amount: Optional[float] = Field(default=None, gt=0, allow_inf_nan=False)
+    amount: Optional[Decimal] = Field(default=None, gt=0, allow_inf_nan=False)
     note: Optional[str] = None
