+++ backend/models/settlement.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import Literal, Optional
 
 from pydantic import BaseModel, Field
@@ -8,7 +9,7 @@ class SettleIn(BaseModel):
     # Kept for backward compatibility; the doc it creates is now stamped status:"paid".
     from_member_id: str
     to_member_id: str
-    amount: float = Field(gt=0, allow_inf_nan=False)
+    amount: Decimal = Field(gt=0, allow_inf_nan=False)
 
 
 class SettlementCreate(BaseModel):
@@ -16,7 +17,7 @@ class SettlementCreate(BaseModel):
     # `status` is server-controlled and always starts "pending" — it is not accepted here.
     from_member_id: str
     to_member_id: str
-    amount: float = Field(gt=0, allow_inf_nan=False)
+    amount: Decimal = Field(gt=0, allow_inf_nan=False)
     note: Optional[str] = None
